requirements.txt
# Core Computer Vision Libraries
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# YOLO and Object Detection
ultralytics>=8.1.0
supervision>=0.20.0

# PDF Processing
PyMuPDF>=1.23.0

# LLM Integration (choose one based on your provider)
# openai>=1.0.0  # For OpenAI GPT models
# anthropic>=0.25.0  # For Anthropic Claude models

# Standard Library Enhancements (already included in Python 3.8+)
# typing, asyncio, concurrent.futures, base64, io, json, dataclasses, os, time, pathlib

# Optional: For enhanced color processing
# scikit-learn>=1.3.0  # If using advanced clustering
# matplotlib>=3.7.0  # For visualization and debugging